Rename: Query와 Command를 분리
- 기존의 OrderService를 QueryOrderService와 CommandOrderService로 분리
- 관련된 클래스에 변경 사항 반영
