+++ src/main/java/com/jelly/zzirit/domain/order/controller/OrderController.java
@@ -2,7 +2,7 @@ package com.jelly.zzirit.domain.order.controller;
 
 import com.jelly.zzirit.domain.order.dto.request.OrderCreateRequest;
 import com.jelly.zzirit.domain.order.dto.response.OrderFetchResponse;
-import com.jelly.zzirit.domain.order.service.OrderService;
+import com.jelly.zzirit.domain.order.service.QueryOrderService;
 import com.jelly.zzirit.global.dto.BaseResponse;
 import com.jelly.zzirit.global.dto.Empty;
 import com.jelly.zzirit.global.security.model.MemberPrincipal;
@@ -20,7 +20,7 @@ import java.util.List;
 @Tag(name = "주문 API", description = "주문과 관련된 API를 설명합니다.")
 public class OrderController {
 
-    private final OrderService orderService;
+    private final QueryOrderService queryOrderService;
 
     @PostMapping
     @Operation(summary = "주문 생성 API", description = "주문 하나를 생성합니다.")
@@ -31,7 +31,7 @@ public class OrderController {
     @GetMapping
     @Operation(summary = "주문 전체 조회 API", description = "전체 주문을 조회합니다.")
     public BaseResponse<List<OrderFetchResponse>> fetchAllOrders(@AuthenticationPrincipal MemberPrincipal member) {
-        List<OrderFetchResponse> response = orderService.findAllOrders(member.getMemberId())
+        List<OrderFetchResponse> response = queryOrderService.findAllOrders(member.getMemberId())
             .stream().map(OrderFetchResponse::from).toList();
 
         return BaseResponse.success(response);
